pyproject.toml
[tool.poetry]
name = "lms-project"
version = "0.1.0"
description = ""
authors = ["Alexandr Zibinin <zibinin.chel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
pillow = "^10.4.0"
black = "^24.4.2"
flake8 = "^7.1.0"
django-filter = "^24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
